Changed Project paths from a static list to an expandable dict.

+++ tofisca/project.py
@@ -3,13 +3,14 @@ from __future__ import annotations
 import shutil
 from enum import Enum
 from pathlib import Path
+from string import Template
 from typing import Union
 
 from pydantic import Field
 
 from app import App
 from configuration.config_item import ProjectItem
-from errors import ProjectAlreadyExistsError, ProjectNotLoadedError
+from errors import ProjectAlreadyExistsError, ProjectNotLoadedError, ProjectDoesNotExistError
 from film_specs import FilmFormat
 from scanarea_manager import ScanAreaManager
 
@@ -31,10 +32,15 @@ class ProjectState(ProjectItem):
     errors: list[str] = Field(default=[])
 
 
-class ProjectPaths(ProjectItem):
-    project_path: str = "{project.name}"
-    scanned_images: str = "scanned_images"
-    final_images: str = "final_images"
+class ProjectPathEntry(ProjectItem):
+    name: str = Field(
+        description="An identifier for this path, e.g. 'project' or 'scanned'")
+    description: str = Field(default="",
+                             description="What the path is for.")
+    path: str = Field(default="",
+                      description="The actual path. May be relative to the application storage folder or absolute. Can contain templates.")
+    resolved: str = Field(default="",
+                          description="The computed absolute path on the filesystem. For info only.")
 
 
 class FilmData(ProjectItem):
@@ -56,6 +62,10 @@ class Project:
 
     @classmethod
     async def load_project(cls, app: App, pid: int) -> Project:
+        """
+        Load the project with the given pid from the config database.
+        :raises: ProjectDoesNotExistError if the project does not exist
+        """
         project = Project(app, pid)
         await project.load()
         return project
@@ -72,7 +82,23 @@ class Project:
 
         self._project_state: ProjectState = ProjectState(pid=pid)
 
-        self._paths: ProjectPaths = ProjectPaths(pid=pid)
+        self._paths: dict[str, ProjectPathEntry] = {
+            "project": ProjectPathEntry(pid=pid,
+                                        name="project",
+                                        description="General project data storage",
+                                        path="${name}",
+                                        ),
+            "scanned": ProjectPathEntry(pid=pid,
+                                        name="scanned",
+                                        description="Folder for raw scanned images",
+                                        path="${project}/scanned_images",
+                                        ),
+            "final": ProjectPathEntry(pid=pid,
+                                      name="final",
+                                      description="Images after processing",
+                                      path="${project}/final_images",
+                                      ),
+        }
 
         self._film_data: FilmData = FilmData(pid=pid)
 
@@ -85,13 +111,18 @@ class Project:
 
         There is no save function.
         All settings are saved whenever they are changed.
+        :raises: ProjectDoesNotExistError if the project with the current pid does not exist.
         """
         self._name = await self.db.get_project_name(self._pid)
+        if self._name is None:
+            raise ProjectDoesNotExistError(self._pid)
+
         await self._project_state.retrieve(self.db, self._pid)
         await self._film_data.retrieve(self.db, self._pid)
         await self._scanarea.load_current_state(self.db, self._pid)
 
-        await self._paths.retrieve(self.db, self._pid)
+        for path in self._paths.values():
+            await path.retrieve(self.db, self._pid)
 
         return self
 
@@ -112,19 +143,22 @@ class Project:
         return self._name
 
     @property
-    def paths(self) -> ProjectPaths:
+    def all_paths(self) -> dict[str, ProjectPathEntry]:
         """
         The paths where to store the images and other data.
 
-        The returned paths may contain templates ('{project.name}' or '{project.id}').
+        The returned paths may contain templates (e.g. '{project.name}' or '{project.id}').
         Use :meth:`resolve_path` to resolve the templates and get the real path.
 
-        This property is read-only and returns a copy of the internal one.
+        This property is read-only.
         Use :meth:`set_paths` to set the project paths.
         """
-        # apply templates
-        paths = self._paths.model_copy()
-        return paths
+        # resolve all paths first before handing them out
+        result: dict[str, ProjectPathEntry] = {}
+        for path_entry in self._paths.values():
+            path_entry.resolved = str(self.resolve_path(path_entry.path, create_folder=False))
+            result[path_entry.name] = path_entry.model_copy()
+        return result
 
     async def set_name(self, new_name: str) -> None:
         """
@@ -147,34 +181,33 @@ class Project:
 
         # change the name in the database
         await self.db.change_project_name(self._pid, new_name)
-        self._name.name = new_name
+        self._name = new_name
 
         # todo: maybe we need to change the name of the paths
 
-    async def set_paths(self, new_paths: ProjectPaths):
+    async def update_path(self, new_path: ProjectPathEntry):
         """
-        Set the image and data storage paths for this project.
+        Update the given path.
         """
-        self._paths.project_path = str(self._update_path(self._paths.project_path, new_paths.project_path))
-        self._paths.scanned_images = str(self._update_path(self._paths.scanned_images, new_paths.scanned_images))
-        self._paths.final_images = str(self._update_path(self._paths.final_images, new_paths.final_images))
+        # get the current path entry
+        try:
+            old_path_entry = self._paths[new_path.name]
+        except KeyError:
+            raise ValueError(f"The Project has no path named {new_path.name}")
 
-        await self._paths.store(self.db, self._pid)
+        old_path_resolved = self.resolve_path(old_path_entry.path, create_folder=False)
 
-    def _update_path(self, old_path: str, new_path: str) -> Path:
-        old_path = self._get_absolute_path(old_path)
-        new_path = self._get_absolute_path(new_path)
-        if new_path != old_path:
-            # project path has changed. rename the folder as well
-            old_path.rename(new_path)
-        return Path(new_path)
+        new_path = new_path.path
+        new_path_resolved = self.resolve_path(new_path, create_folder=False)
 
-    def _get_absolute_path(self, path: Path | str) -> Path:
-        p = Path(path)
-        if p.is_absolute():
-            return p
-        else:
-            return self.app.project_manager.root_path / path
+        if new_path_resolved != old_path_resolved and old_path_resolved.exists():
+            # The Path has changed. Rename the folder as well (if it exists)
+            old_path_resolved.rename(new_path_resolved)
+
+        # update the database
+        self._paths[old_path_entry.name].path = new_path
+        self._paths[old_path_entry.name].resolved = new_path_resolved
+        await self._paths[old_path_entry.name].store(self.db, self._pid)
 
     async def _delete_storage(self) -> None:
         """
@@ -184,14 +217,10 @@ class Project:
 
         :raises Exception: A filesystem Extecption if the storage paths could not be deleted.
         """
-        path = self.resolve_path(self._paths.project_path)
-        shutil.rmtree(path)
-
-        path = self.resolve_path(self._paths.scanned_images)
-        shutil.rmtree(path)
-
-        path = self.resolve_path(self._paths.final_images)
-        shutil.rmtree(path)
+        for entry in self._paths.values():
+            path = self.resolve_path(entry.path, create_folder=False)
+            if path.exists():
+                shutil.rmtree(path)
 
     def resolve_path(self, folder_path: str | Path, create_folder: bool = True) -> Path:
         """
@@ -211,8 +240,20 @@ class Project:
         :raises Exception: A filesystem Extecption if the storage paths could not be created.
         """
         folder_path = str(folder_path)  # convert Path to string...
-        folder_path = folder_path.replace("{project.name}", self._name)
-        folder_path = folder_path.replace("{project.id}", str(self._pid))
+
+        # first build a list of all template identifiers
+        identifiers = {"name": self._name, "pid": self._pid, }
+        for key in self._paths.keys():
+            identifiers[key] = self._paths[key].path
+
+        # now substitude until no more substitutions left
+        counter = 0
+        while "${" in folder_path:
+            resolver = Template(folder_path)
+            folder_path = resolver.safe_substitute(identifiers)
+            counter += 1
+            if counter > 10:
+                raise RuntimeError("Template substitution counter exceeded. Probably due to circular template.")
 
         folder_path = Path(folder_path)  # ... and back to Path
         if not folder_path.is_absolute():
